Первая версия сервисов лабы 2

+++ src/main/java/ru/mirea/webPetShop/myPets/ServiceElements/CartElement.java
@@ -1,25 +1,25 @@
 package ru.mirea.webPetShop.myPets.ServiceElements;
 
 public class CartElement {
-    private int id;
-    private int person_id;
-    private int item_id;
+    private long id;
+    private long person_id;
+    private long item_id;
 
-    public CartElement(int id, int person_id, int item_id) {
+    public CartElement(long id, long person_id, long item_id) {
         this.id = id;
         this.person_id = person_id;
         this.item_id = item_id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public int getPerson_id() {
+    public long getPerson_id() {
         return person_id;
     }
 
-    public int getItem_id() {
+    public long getItem_id() {
         return item_id;
     }
 }
